Store nested trees in the object database Currently, each commit contains only a single tree that contains a flat list of blobs. To store projects with subdirectories, we need to write nested trees to the database.
These paths are stored off as blobs and wrapped in `Entry` objects in
the usual way. This list is then turned into a nested structure by
`Tree.Build`, and we scan this structure using `Tree#traverse`
to store off all the subtrees. Saving each subtree gives it an object ID
that allows it to be listed in its parent tree.