Discard index files whose directories collide with files If the index contains the following files:
    alice.txt
    nested/bob.txt
    nested/inner/claire.txt

and a file named `nested` is added, then the entries `nested/bob.txt`
and `nested/inner/claire.txt` should be removed, since the name `nested`
now corresponds to a file, not a directory.

By adding a new structure to `Index` that tracks all the entries that
are downstream of a particular directory, we make it easy to identify
entries that must be removed when one of their parent directories is
replaced by a file.