Introduce the Repository class The `add` and `commit` commands both begin with a lot of similar setup that defines the paths to various pieces of Git infrastructure and then creates objects like `Database` and `Index` around them.
Here we extract this common setup into a single class that knows how to
locate each component on the filesystem and instantiate its class. This
gets rid of some boilerplate from the command code and reduces the code
we'll need to write in each new command we add.

The methods in `Repository` are memoised, because some of the objects it
returns are stateful, like `Index`. Therefore, we'd like every call to
`Repository#index` to return the same object.